refactor(trackers): move validate_supported_dataset
Move validate_supported_dataset function from tracker_stock to validation module.
